N->D   이상한 점 발견. BoxCollider를 Orthographic으로 렌더시, 잘 된다고 생각했는데 카메라 돌리면 그냥 UI처럼 화면중앙에서 돌기만함. 애초에 처음부터 잘못됐었다.

+++ DX113D_2004/Environment/Environment.cpp
@@ -169,7 +169,7 @@ void Environment::createPerspectiveBuffer()
 
 void Environment::createOrthographicBuffer()
 {
-	mOrthographicProjectionMatrix = XMMatrixOrthographicLH(WIN_WIDTH, WIN_HEIGHT, 0.1f, 1000.0f);
+	mOrthographicProjectionMatrix = XMMatrixOrthographicLH((float)WIN_WIDTH, (float)WIN_HEIGHT, 0.1f, 1000.0f);
 
 	mOrthographicProjectionBuffer = new ProjectionBuffer();
 	mOrthographicProjectionBuffer->Set(mOrthographicProjectionMatrix);
